requirements.txt
django > 4.2
djangorestframework
psycopg2
djangorestframework-simplejwt